requirements.txt
# streamlit
# google-generativeai
# PyMuPDF
# python-dotenv

streamlit>=1.28.0
google-generativeai>=0.3.2
PyMuPDF>=1.23.0
python-dotenv>=1.0.0